add fill array with values
